pyproject.toml
[tool.poetry]
name = "aitestdrive"
version = "0.1.0"
description = ""
authors = ["etvt <149910696+etvt@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
google-cloud-aiplatform = "^1.36.4"
google-cloud-storage = "^2.13.0"
qdrant-client = "^1.6.9"
langchain = "^0.0.347"
tiktoken = "^0.5.2"
pdfplumber = "^0.10.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "aitestdrive.main:main"
server_dev = "aitestdrive.main:main_dev"